+++ lib/address_picker/route/address_picker_route.dart
@@ -53,12 +53,14 @@ class AddressPickerRoute<T> extends PopupRoute<T> {
   @override
   AnimationController createAnimationController() {
     assert(_animationController == null);
-    _animationController = BottomSheet.createAnimationController(navigator!.overlay!);
+    _animationController =
+        BottomSheet.createAnimationController(navigator!.overlay!);
     return _animationController!;
   }
 
   @override
-  Widget buildPage(BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation) {
+  Widget buildPage(BuildContext context, Animation<double> animation,
+      Animation<double> secondaryAnimation) {
     Widget bottomSheet = MediaQuery.removePadding(
       context: context,
       removeTop: true,
@@ -97,8 +99,8 @@ class _PickerContentView extends StatefulWidget {
   final PickerStyle pickerStyle;
 
   @override
-  State<StatefulWidget> createState() =>
-      _PickerState(this.initProvince, this.initCity, this.initTown, this.addAllItem, this.pickerStyle);
+  State<StatefulWidget> createState() => _PickerState(this.initProvince,
+      this.initCity, this.initTown, this.addAllItem, this.pickerStyle);
 }
 
 class _PickerState extends State<_PickerContentView> {
@@ -118,9 +120,12 @@ class _PickerState extends State<_PickerContentView> {
   AnimationController? controller;
   Animation<double>? animation;
 
-  late FixedExtentScrollController provinceScrollCtrl, cityScrollCtrl, townScrollCtrl;
+  late FixedExtentScrollController provinceScrollCtrl,
+      cityScrollCtrl,
+      townScrollCtrl;
 
-  _PickerState(this._currentProvince, this._currentCity, this._currentTown, this.addAllItem, this._pickerStyle) {
+  _PickerState(this._currentProvince, this._currentCity, this._currentTown,
+      this.addAllItem, this._pickerStyle) {
     provinces = Address.provinces;
     hasTown = this._currentTown != null;
 
@@ -144,7 +149,8 @@ class _PickerState extends State<_PickerContentView> {
         builder: (BuildContext context, Widget? child) {
           return ClipRect(
             child: CustomSingleChildLayout(
-              delegate: _BottomPickerLayout(widget.route.animation!.value, this._pickerStyle),
+              delegate: _BottomPickerLayout(
+                  widget.route.animation!.value, this._pickerStyle),
               child: GestureDetector(
                 child: Material(
                   color: Colors.transparent,
@@ -292,6 +298,7 @@ class _PickerState extends State<_PickerContentView> {
               padding: const EdgeInsets.all(8.0),
               child: CupertinoPicker.builder(
                 scrollController: provinceScrollCtrl,
+                selectionOverlay: _pickerStyle.itemOverlay,
                 itemExtent: _pickerStyle.pickerItemHeight,
                 onSelectedItemChanged: (int index) {
                   _setProvince(index);
@@ -302,7 +309,9 @@ class _PickerState extends State<_PickerContentView> {
                   return Align(
                       alignment: Alignment.center,
                       child: Text(text,
-                          style: TextStyle(color: _pickerStyle.textColor, fontSize: _pickerFontSize(text)),
+                          style: TextStyle(
+                              color: _pickerStyle.textColor,
+                              fontSize: _pickerFontSize(text)),
                           textAlign: TextAlign.start));
                 },
               ),
@@ -313,6 +322,7 @@ class _PickerState extends State<_PickerContentView> {
                 padding: EdgeInsets.all(8.0),
                 child: CupertinoPicker.builder(
                   scrollController: cityScrollCtrl,
+                  selectionOverlay: _pickerStyle.itemOverlay,
                   itemExtent: _pickerStyle.pickerItemHeight,
                   onSelectedItemChanged: (int index) {
                     _setCity(index);
@@ -323,7 +333,9 @@ class _PickerState extends State<_PickerContentView> {
                     return Align(
                       alignment: Alignment.center,
                       child: Text('$text',
-                          style: TextStyle(color: _pickerStyle.textColor, fontSize: _pickerFontSize(text)),
+                          style: TextStyle(
+                              color: _pickerStyle.textColor,
+                              fontSize: _pickerFontSize(text)),
                           textAlign: TextAlign.start),
                     );
                   },
@@ -335,6 +347,7 @@ class _PickerState extends State<_PickerContentView> {
                       padding: EdgeInsets.all(8.0),
                       child: CupertinoPicker.builder(
                         scrollController: townScrollCtrl,
+                        selectionOverlay: _pickerStyle.itemOverlay,
                         itemExtent: _pickerStyle.pickerItemHeight,
                         onSelectedItemChanged: (int index) {
                           _setTown(index);
@@ -345,7 +358,9 @@ class _PickerState extends State<_PickerContentView> {
                           return Align(
                             alignment: Alignment.center,
                             child: Text(text,
-                                style: TextStyle(color: _pickerStyle.textColor, fontSize: _pickerFontSize(text)),
+                                style: TextStyle(
+                                    color: _pickerStyle.textColor,
+                                    fontSize: _pickerFontSize(text)),
                                 textAlign: TextAlign.start),
                           );
                         },
@@ -367,7 +382,9 @@ class _PickerState extends State<_PickerContentView> {
         crossAxisAlignment: CrossAxisAlignment.center,
         children: <Widget>[
           /// 取消按钮
-          InkWell(onTap: () => Navigator.pop(context), child: _pickerStyle.cancelButton),
+          InkWell(
+              onTap: () => Navigator.pop(context),
+              child: _pickerStyle.cancelButton),
 
           /// 标题
           Expanded(child: _pickerStyle.title),
@@ -376,7 +393,8 @@ class _PickerState extends State<_PickerContentView> {
           InkWell(
               onTap: () {
                 if (widget.route.onConfirm != null) {
-                  widget.route.onConfirm!(_currentProvince, _currentCity, _currentTown);
+                  widget.route.onConfirm!(
+                      _currentProvince, _currentCity, _currentTown);
                 }
                 Navigator.pop(context);
               },
@@ -404,7 +422,10 @@ class _BottomPickerLayout extends SingleChildLayoutDelegate {
     }
 
     return BoxConstraints(
-        minWidth: constraints.maxWidth, maxWidth: constraints.maxWidth, minHeight: 0.0, maxHeight: maxHeight);
+        minWidth: constraints.maxWidth,
+        maxWidth: constraints.maxWidth,
+        minHeight: 0.0,
+        maxHeight: maxHeight);
   }
 
   @override